+++ src/gen/opc/core.ts
@@ -6,7 +6,22 @@
  */
 
 import { XML_DECL } from '../../core-enums.js'
-import { encodeXmlEntities } from '../../gen-utils.js'
+import { el, raw } from '../oxml/el.js'
+
+/** Each property sits on its own indented line; the parent supplies the closing indent. */
+const PROP = { openPrefix: '\n\t\t' }
+
+const NS = {
+	'xmlns:cp': 'http://schemas.openxmlformats.org/package/2006/metadata/core-properties',
+	'xmlns:dc': 'http://purl.org/dc/elements/1.1/',
+	'xmlns:dcterms': 'http://purl.org/dc/terms/',
+	'xmlns:dcmitype': 'http://purl.org/dc/dcmitype/',
+	'xmlns:xsi': 'http://www.w3.org/2001/XMLSchema-instance',
+}
+
+function timestamp(): string {
+	return new Date().toISOString().replace(/\.\d\d\dZ/, 'Z')
+}
 
 /**
  * Creates `docProps/core.xml`
@@ -17,14 +32,23 @@ import { encodeXmlEntities } from '../../gen-utils.js'
  * @returns XML
  */
 export function makeXmlCore(title: string, subject: string, author: string, revision: string): string {
-	return `${XML_DECL}
-	<cp:coreProperties xmlns:cp="http://schemas.openxmlformats.org/package/2006/metadata/core-properties" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:dcmitype="http://purl.org/dc/dcmitype/" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-		<dc:title>${encodeXmlEntities(title)}</dc:title>
-		<dc:subject>${encodeXmlEntities(subject)}</dc:subject>
-		<dc:creator>${encodeXmlEntities(author)}</dc:creator>
-		<cp:lastModifiedBy>${encodeXmlEntities(author)}</cp:lastModifiedBy>
-		<cp:revision>${revision}</cp:revision>
-		<dcterms:created xsi:type="dcterms:W3CDTF">${new Date().toISOString().replace(/\.\d\d\dZ/, 'Z')}</dcterms:created>
-		<dcterms:modified xsi:type="dcterms:W3CDTF">${new Date().toISOString().replace(/\.\d\d\dZ/, 'Z')}</dcterms:modified>
-	</cp:coreProperties>`
+	const dcterms = { 'xsi:type': 'dcterms:W3CDTF' }
+	return (
+		XML_DECL +
+		el(
+			'cp:coreProperties',
+			NS,
+			[
+				raw(el('dc:title', null, title, PROP)),
+				raw(el('dc:subject', null, subject, PROP)),
+				raw(el('dc:creator', null, author, PROP)),
+				raw(el('cp:lastModifiedBy', null, author, PROP)),
+				// `revision` is interpolated unescaped today; raw() preserves that.
+				raw(el('cp:revision', null, raw(revision), PROP)),
+				raw(el('dcterms:created', dcterms, raw(timestamp()), PROP)),
+				raw(el('dcterms:modified', dcterms, raw(timestamp()), PROP)),
+			],
+			{ openPrefix: '\n\t', closePrefix: '\n\t' }
+		)
+	)
 }